* added comments in the sdk
